gerstner, grid: moved GridHeight class to GridData class
GridHeght was a 2D array of ints implemented in gerstner.h.  This was abstracted to template<typename T> class GridData to allow arbitrary collections of data to be associated with each grid point.
This is in preparation for long-term separation of gridlib from gerstner, allowing its use in other projects.