Gerst4: Gerst3 moved to LEDMatrixWcs
1. Moved Gerst3 to LEDMatrixWcs libraries.
2. Updated some of the debugging code.
Note: This is an interim release to test integration with the LEDMatrixWcs libraries.  It is not meant for release.  Gerstner5 will have a full conversion to FP.
Note:  FPS rate dropped from 88 to 83, but that may be due to some of the added FP/Int conversions.
Note: The color peaks aren't hitting as well as before.  We'll need to retune the color mapping code (for a later release)